# Conflicts:
#   .gitlab-ci.yml
#   README.md
#   build.gradle
#   gradle.properties
#   src/commonMain/kotlin/com/jessecorbett/diskord/api/rest/client/ChannelClient.kt
#   src/commonMain/kotlin/com/jessecorbett/diskord/internal/CommonResources.kt

+++ src/commonMain/kotlin/com/jessecorbett/diskord/api/websocket/WebSocketCloseCode.kt
@@ -5,7 +5,7 @@ package com.jessecorbett.diskord.api.websocket
  *
  * @property code the close code represented.
  */
-enum class WebSocketCloseCode(val code: Int) {
+enum class WebSocketCloseCode(val code: Short) {
     /*
      * Standard Socket Codes
      * https://tools.ietf.org/html/rfc6455#section-7.4
